pyproject.toml
[project]
name = "loco-operator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "python-dotenv",
]
